Add files via upload
Updated to use Gaia DR3. Various features have been tweaked or added. User can now maintain a local file (LocalRV.csv) to lookup extra RVs for use in the analysis. The RV handling also bakes in the hot-star RV bias correction. Given the addition of faint star RVs with relatively large errors, the threshold for "nonagreement" in the figures now requires an RV discrepancy that is larger than the user-specified RVcut and larger than 2 sigma. The threshold for agreement is still only to be smaller than RVcut.

As a bonus, there is also an alpha version of my "Binary Probability Calculator", called as Comove.binprob. This routine computes the probability that a candidate companion to a star is either a bound companion or a background star. This is based in the math laid out by Deacon et al. (2016, 2020). I make no guarantees as to its operation at this point, but feedback is welcome.